# spec/openapi.yaml
openapi: 3.1.0
info:
  title: PAL NAS Media Uploader API
  version: 1.0.0
  description: |
    Backend API for a self-hosted, multi-user NAS media library.

    ## Accounts and authorization

    The service has two roles. Administrators manage accounts but cannot use
    upload or gallery routes. Regular users upload into their configured folder,
    organize accessible media, and share owned media with other regular users.
    Resource lookup is scoped by the authenticated identity; inaccessible
    resources generally return `404` instead of revealing their existence.

    The first persistent administrator is created through the one-time HTML
    setup flow. Deployment bootstrap credentials authorize that operation only
    and are never stored as an application account. The initial administrator
    enrolls TOTP on first login. Every account subsequently created by an
    administrator also uses password plus TOTP.

    ## Sessions and browser access

    Authentication produces an opaque, HttpOnly cookie named
    `pal_medias_uploader_session`. Sessions last 30 days. Cross-origin browser
    requests must send credentials and their `Origin` must appear in
    `ENV_CORS_ALLOWED_ORIGINS`; an unlisted origin receives an empty `403`
    response. The cookie is `Secure` for HTTPS requests and uses `SameSite=None`
    only for secure cross-site requests; otherwise it uses `SameSite=Strict`.

    ## Uploads and media

...

...

    ## Albums

...
    have membership changed manually.

...

...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /admin/operations/integrity:
    get:
      tags: [Operations]
      operationId: scanIntegrity
      summary: Compare media records with filesystem files
      responses:
        "200":
          description: Integrity report.
          content:
            application/json:
              schema:
                type: object
                required: [filesWithoutRecords, recordsWithoutFiles, checkedAt]
                properties:
                  filesWithoutRecords:
                    type: array
                    items: { type: string }
                  recordsWithoutFiles:
                    type: array
                    items: { type: string }
                  checkedAt:
                    type: string
                    format: date-time
        "403":
          description: Administrator role required.
        "500":
          description: Integrity scan failed.
...
  /auth/login:
    post:
      tags: [Authentication]
      operationId: beginLogin
      summary: Log in with a password and optional TOTP
      description: |
        Accepts a JSON body of at most 1 MiB and rejects unknown fields. The
        username is trimmed and matched case-sensitively; the password is checked
        with bcrypt.

        The initial configured administrator receives a session cookie
        immediately and the response contains `authenticated: true`. Every
        subsequently created account receives an opaque challenge valid for five
        minutes. Managed accounts receive their TOTP provisioning URI when an
        administrator creates them. The login response exposes provisioning data
        only if an account has a pending, not-yet-active TOTP secret.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: A TOTP challenge was created, or the initial administrator was authenticated immediately.
          headers:
            Set-Cookie:
              description: Set immediately for accounts exempt from TOTP.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginChallenge"
              examples:
                initialAdministrator:
                  summary: Password-only initial administrator
                  value:
                    authenticated: true
                totpChallenge:
                  summary: Existing TOTP-enabled account
                  value:
                    challengeToken: opaque-challenge-token
                enrollmentChallenge:
                  summary: First login for a managed account
                  value:
                    challengeToken: opaque-challenge-token
                    provisioningUri: otpauth://totp/PAL:user
                    secret: JBSWY3DPEHPK3PXP
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrigin"
        "409":
          $ref: "#/components/responses/UploadConflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/verify:
    post:
      tags: [Authentication]
      operationId: verifyTOTP
      summary: Verify TOTP and create a session
      description: |
        Validates an unexpired challenge and a six-digit TOTP code. TOTP uses a
        30-second period, SHA-1, six digits, and accepts one adjacent time step in
        either direction. For first-time login, successful verification promotes
        the pending secret to the account's active secret. The challenge is then
        deleted and cannot be reused.

        A successful response creates a server-side session valid for 30 days
        and sets the opaque HttpOnly session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTOTPRequest"
      responses:
        "200":
          description: Authenticated session created.
          headers:
            Set-Cookie:
              description: HttpOnly session cookie named pal_medias_uploader_session.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrigin"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/session:
    get:
      tags: [Authentication]
      operationId: getSession
      summary: Get the current session
      description: Resolves the opaque session cookie and returns the current account. Expired, missing, or unknown sessions receive `401`.
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrigin"
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Delete the current session
      description: Requires a valid session, deletes it from the server, and expires the browser cookie. Reusing the deleted cookie receives `401`.
      responses:
        "204":
          description: Session deleted and cookie expired.
          headers:
            Set-Cookie:
              description: Expired pal_medias_uploader_session cookie.
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrigin"
  /admin/config:
    get:
      tags: [Users]
      operationId: userConfigurationPage
      summary: Open the administrator configuration page
      description: Requires an administrator session. The embedded HTML page calls `/admin/users`; disabling frontend pages does not disable those JSON endpoints.
      responses:
        "200":
          description: User configuration HTML page.
          content:
            text/html:
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Frontend pages are disabled.
  /admin/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      description: Returns every configured account ordered by username. `totpEnabled` reflects whether an active TOTP secret is stored.
      responses:
        "200":
          description: All configured users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a TOTP-enabled user
      description: |
        Accepts a JSON body of at most 1 MiB and rejects unknown fields. The
        username is trimmed and case-sensitive. Roles are trimmed and normalized
        to lowercase. Passwords are bcrypt-hashed.

        A regular user requires a non-empty relative upload folder. Nested paths
        such as `family/alice` are allowed, while absolute paths and parent
        traversal are rejected. Administrator folders are always cleared.

        The service generates a TOTP secret for every managed account, including
        additional administrators. The provisioning URI and PNG data URI are
        returned only in this creation response and should be delivered securely.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created. The TOTP QR code is shown only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/me/folder:
    put:
      tags: [Users]
      operationId: setMyUploadFolder
      summary: Set the current regular user's upload folder
      description: |
        Available only to role `user`. The folder is trimmed, cleaned, and stored
        with forward slashes. It must be non-empty and relative, and cannot be
        `..` or begin with `../`. Changing it affects the destination of future
        completed uploads; it does not move existing media.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [uploadFolder]
              properties:
                uploadFolder:
                  type: string
                  examples: [family/alice]
      responses:
        "204":
          description: Folder updated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /media/files/{id}/download:
    get:
      tags: [Media]
      operationId: downloadMedia
      summary: Download owned or shared media
      description: Returns an active media file only when the current user owns it or has a share. Filesystem paths are constrained to the configured media root.
      parameters:
        - $ref: "#/components/parameters/UploadID"
      responses:
        "200":
          description: Media bytes.
          headers:
            Content-Disposition:
              description: Attachment disposition containing the stored filename.
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/UploadNotFound"
  /media/files/{id}/shares:
    parameters:
      - $ref: "#/components/parameters/UploadID"
    post:
      tags: [Media]
      operationId: shareMedia
      summary: Share owned media with a regular user
      description: |
        Creates a library-wide share from the current user to the target user.
        The path media ID is used to verify that the current user owns active
        media; the grant is not limited to that item. The recipient immediately
        gains access to all of the owner's current active media, and future
        uploads become accessible without another share request.

        Shared media contributes to the recipient's lists, downloads, manual
        albums, favorites, storage totals, and next nightly automatic-album
        reconciliation. The target must be another regular user; administrators
        and the owner cannot be recipients. Repeating the relationship is
        idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareRequest"
      responses:
        "204":
          description: Library share added or already present.
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      tags: [Media]
      operationId: unshareMedia
      summary: Remove a library share
      description: |
        Removes the owner-to-recipient relationship represented by the owned
        media ID. All media belonging to that owner immediately becomes
        inaccessible to the recipient. It disappears from lists, downloads,
        storage totals, and album responses; future uploads are no longer shared.
        Ownership and the owner's albums are unchanged. Legacy clients may still
        send the username JSON body, but new clients should use the query
        parameter.
      parameters:
        - name: username
          in: query
          required: true
          description: Case-sensitive username of the regular-user recipient to remove.
          schema:
            type: string
            minLength: 1
      responses:
        "204":
          description: Library share removed.
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /auth/recovery:
    post:
      tags: [Authentication]
      operationId: verifyRecoveryCode
      summary: Authenticate with a single-use TOTP recovery code
      description: Requires the short-lived challenge returned after successful password verification. Three invalid attempts are rate limited with the same controls as TOTP verification.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryVerificationRequest"
      responses:
        "200":
          description: Recovery code consumed and authenticated session created.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: Too many authentication attempts.
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/recovery-codes:
    post:
      tags: [Authentication]
      operationId: regenerateRecoveryCodes
      summary: Replace and return TOTP recovery codes
      description: Invalidates every previous code and returns ten new single-use codes. Only hashes are retained; this response is the only opportunity to save the plaintext values.
      responses:
        "201":
          description: New recovery-code set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryCodesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /admin/users/{id}/password:
    patch:
      tags: [Users]
      operationId: resetUserPassword
      summary: Reset a user's password and revoke their sessions
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "204":
          description: Password replaced and all target-user sessions deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}/quota:
    patch:
      tags: [Users]
      operationId: setUserStorageQuota
      summary: Set or clear a per-user storage quota
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotaRequest"
      responses:
        "204":
          description: Quota updated. A null value restores the environment default.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}/trash-retention:
    patch:
      tags: [Users]
      operationId: setUserTrashRetention
      summary: Set a user's trash retention period
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrashRetentionRequest"
      responses:
        "204":
          description: Retention period updated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/storage/dashboard:
    get:
      tags: [Users]
      operationId: getStorageDashboard
      summary: Get aggregate and per-user storage usage
      responses:
        "200":
          description: Storage dashboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageDashboard"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/processing-jobs:
    get:
      tags: [Processing]
      operationId: listProcessingJobs
      summary: List the 200 most recently updated media-processing jobs
      responses:
        "200":
          description: Processing jobs ordered by most recent update.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcessingJob"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /admin/processing-jobs/{id}/retry:
    post:
      tags: [Processing]
      operationId: retryProcessingJob
      summary: Retry a failed processing job
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "204":
          description: Job reset to queued with its attempt count cleared.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /admin/trash/cleanup:
    post:
      tags: [Operations]
      operationId: cleanupExpiredTrash
      summary: Permanently remove expired trash immediately
      responses:
        "200":
          description: Cleanup result for up to 500 records.
          content:
            application/json:
              schema:
                type: object
                required: [removed]
                properties:
                  removed: { type: integer, minimum: 0 }
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /media/app:
    get:
      tags: [Media]
      operationId: mediaWorkspacePage
      summary: Open the embedded media workspace
      responses:
        "200":
          description: Authenticated media workspace HTML.
          content:
            text/html:
              schema: { type: string }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Frontend pages are disabled.
...
  /media/files/{id}/thumbnail:
    get:
      tags: [Media]
      operationId: downloadThumbnail
      summary: Get a generated thumbnail for accessible media
      parameters:
        - $ref: "#/components/parameters/UploadID"
      responses:
        "200":
          description: Generated JPEG thumbnail.
          headers:
            Cache-Control:
              schema: { type: string }
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /media/files/{id}/processing-status:
    get:
      tags: [Processing]
      operationId: getMediaProcessingStatus
      summary: Get processing status for accessible media
      parameters:
        - $ref: "#/components/parameters/UploadID"
      responses:
        "200":
          description: Processing job status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingJob"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
...
  /media/public-links/{id}:
    delete:
      tags: [Public Sharing]
      operationId: revokePublicLink
      summary: Revoke an owned public link
      parameters:
        - $ref: "#/components/parameters/PublicLinkID"
      responses:
        "204":
          description: Public link revoked.
        "404":
          $ref: "#/components/responses/ResourceNotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pal_medias_uploader_session
      description: |
        Opaque session token stored only as a hash by the server. It is issued
        after successful TOTP verification, or immediately for the initial
        administrator, and expires after 30 days. Browser cookies are HttpOnly,
        use path `/`, and are Secure whenever the request is HTTPS.
...